// Features/reports/Types/IAccumulateForReport.ts
export interface IAccumulateForReport {
    identifier: string;
    idAccumulate: number;
    idEmployee: number;
    employeeName: string;
    date: string;
    amount: number;
    idConcept: number;
    conceptCode: string;
    concept: string;
    idPayrollPay: number;
    payrollName: string;
    idCompany: number;
}